[NEW] Add new endpoint for get assessment result

+++ src/modules/assessement/assessment.controller.ts
@@ -25,6 +25,9 @@ export class AssessmentController {
 
     static deleteAssessment = asyncHandler(async (req: Request, res: Response) => {
         const id = Number(req.params.id);
+        if (!id) {
+            throw new Error("ID is required");
+        }
         const result = await AssessmentService.deleteAssessment(id);
         res.status(200).json({
             success: true,
@@ -34,10 +37,26 @@ export class AssessmentController {
 
     static getAssessmentById = asyncHandler(async (req: Request, res: Response) => {
         const id = Number(req.params.id);
+        if (!id) {
+            throw new Error("ID is required");
+        }
         const result = await AssessmentService.getAssessmentById(id);
         res.status(200).json({
             success: true,
             message: "Assessment berhasil diambil",
+            data: result,
+        });
+    });
+
+    static getAssessmentResultDetails = asyncHandler(async (req: Request, res: Response) => {
+        const resultId = Number(req.params.resultId);
+        if (!resultId) {
+            throw new Error("Result ID is required");
+        }
+        const result = await AssessmentService.getAssessmentResultDetails(resultId);
+        res.status(200).json({
+            success: true,
+            message: "Detail hasil assessment berhasil diambil",
             data: result,
         });
     });
